Unify mock socket
