+++ src/test.py
@@ -18,19 +18,11 @@ class socket:
     def __init__(self, *args, **kwargs):
         self.request = b""
         self.connected = False
-        self.scheme = "http"
-        self.ssl_hostname = None
 
     def connect(self, host_port):
+        self.scheme = "http"
         self.host, self.port = host_port
         self.connected = True
-        self._check_cert()
-
-    def _check_cert(self):
-        if self.connected and self.host and self.ssl_hostname:
-            assert self.host == self.ssl_hostname, "server_hostname does not match the host"
-            if self.host.endswith(".badssl.com"): # Fake badssl.com
-                raise ssl.SSLCertVerificationError()
 
     def send(self, text):
         self.request += text
@@ -43,9 +35,8 @@ class socket:
             assert all(int(value) == len(self.body) for name, value in headers
                        if name.lower() == "content-length")
 
-    def makefile(self, mode, encoding, newline):
-        assert self.connected and self.host and self.port, \
-            "You cannot call makefile() on a socket until you call connect() and send()"
+    def makefile(self, mode, encoding=None, newline=None):
+        assert self.connected and self.host and self.port
         if self.port == 80 and self.scheme == "http":
             url = self.scheme + "://" + self.host + self.path
         elif self.port == 443 and self.scheme == "https":
@@ -57,7 +48,14 @@ class socket:
         output = self.URLs[url][1]
         if self.URLs[url][2]:
             assert self.body == self.URLs[url][2], (self.body, self.URLs[url][2])
-        return io.StringIO(output.decode(encoding).replace(newline, "\n"), newline)
+        stream = io.BytesIO(output)
+        if encoding:
+            stream = io.TextIOWrapper(stream, encoding=encoding, newline=newline)
+            stream.mode = mode
+        else:
+            assert mode == "b", "If no file encoding is passed, must pass 'b' mode"
+
+        return stream
 
     def close(self):
         self.connected = False
@@ -73,7 +71,7 @@ class socket:
     @classmethod
     def respond_ok(cls, url, response, method="GET", body=None):
         response = ("HTTP/1.0 200 OK\r\n\r\n" + response).encode("utf8")
-        cls.respond(url, response, method=method, body=body)
+        cls.URLs[url] = [method, response, body]
 
     @classmethod
     def serve(cls, html):
@@ -101,8 +99,7 @@ class socket:
 
 class ssl:
     def wrap_socket(self, s, server_hostname):
-        s.ssl_hostname = server_hostname
-        s._check_cert()
+        assert s.host == server_hostname
         s.scheme = "https"
         return s
